+++ Editor/AvatarProcessor.cs
@@ -84,18 +84,27 @@ namespace nadena.dev.ndmf
             return PlatformExtensions.CanProcessObject(avatar);
         }
 
+        [Obsolete("ProcessAvatarUI() does not handle platforms correctly. Please use ManualProcessAvatar() instead, and specify VRChatPlatform to retain the behavior of ProcessAvatarUI().")]
+        public static GameObject ProcessAvatarUI(GameObject obj)
+        {
+            return ManualProcessAvatar(obj, AmbientPlatform.DefaultPlatform);
+        }
+
         /// <summary>
         /// Process an avatar on request by the user. The resulting assets will be saved in a persistent directory
         /// that will not be cleaned up by CleanTemporaryAssets.
         /// </summary>
         /// <param name="obj"></param>
+        /// <param name="platform"></param>
         /// <returns></returns>
-        public static GameObject ProcessAvatarUI(GameObject obj)
+        [PublicAPI]
+        public static GameObject ManualProcessAvatar(GameObject obj, INDMFPlatformProvider platform = null)
         {
             using (new OverrideTemporaryDirectoryScope("Assets/ZZZ_GeneratedAssets"))
             {
                 var avatar = UnityObject.Instantiate(obj);
-                var buildContext = new BuildContext(avatar, TemporaryAssetRoot);
+                platform ??= AmbientPlatform.CurrentPlatform;
+                var buildContext = new BuildContext(avatar, TemporaryAssetRoot, platform);
 
                 avatar.transform.position += Vector3.forward * 2f;
                 try